udacity/Cloud Computing/Cloud Developer using Microsoft Azure/Migrate App to Azure/function/requirements.txt
azure-functions
psycopg2-binary==2.8.6
sendgrid==6.6.0